implemented all-maps settings
I chose this implementation, because it doesn't require new variables (which caused issues in the past). A bit unclean and only works if the map has exactly 18 holes.

Resets when exiting to main menu, even when the player just finished a map. This could be problematic, when the player accidentally presses "Exit" after finishing a map, when doing 100% category. But its better than nothing for now.